+++ internal/captcha/classification_captcha_manager.go
@@ -1,8 +1,12 @@
 package captcha
 
 import (
+	"bytes"
 	"context"
+	"encoding/base64"
 	"encoding/json"
+	"fmt"
+	"image/jpeg"
 	"math/rand"
 	"os"
 	"slices"
@@ -10,32 +14,38 @@ import (
 	"time"
 )
 
-type CaptchaTask struct {
-	Id          uint32 `json:"id"`
-	Description string `json:"description,omitempty"`
-	Complexity  int    `json:"complexity,omitempty"`
-	Images      []string
-	Solution    []int
-	ts          time.Time
+type ChallengeTemplate struct {
+	Description string   `json:"description,omitempty"`
+	Images      []string `json:"images"`
 }
 
-type CaptchaSolution struct {
+type Challenge struct {
+	ChallengeTemplate
+	Id         uint32 `json:"id"`
+	Complexity int    `json:"complexity,omitempty"`
+	Solution   []int
+	ts         time.Time
+}
+
+type Solution struct {
 	Id       uint32 `json:"id"`
 	Solution []int  `json:"solution,omitempty"`
 }
 
-type templatesConfiguration struct {
-	Templates []CaptchaTask `json:"templates"`
+type Configuration struct {
+	Templates []Challenge `json:"templates"`
 }
 
-type ClassificationCaptchaManager struct {
-	temlates   []CaptchaTask
-	tasks      map[uint32]*CaptchaTask
+type CaptchaManager struct {
+	ctx        context.Context
+	templates  []Challenge
+	tasks      map[uint32]*Challenge
 	complexity int
+	imageBytes map[string]string
 	mu         sync.Mutex
 }
 
-func (c *ClassificationCaptchaManager) cleanup() {
+func (c *CaptchaManager) cleanup() {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	for id, task := range c.tasks {
@@ -45,58 +55,49 @@ func (c *ClassificationCaptchaManager) cleanup() {
 	}
 }
 
-func (c *ClassificationCaptchaManager) Serve(ctx context.Context) (err error) {
-	// Load templates configuration
-	var templates templatesConfiguration
-	content, err := os.ReadFile("/etc/aegis/captcha.json")
-	if err != nil {
-		return
-	}
-	err = json.Unmarshal(content, &templates)
-	if err != nil {
-		return
-	}
-	c.mu.Lock()
-	c.temlates = templates.Templates
-	c.mu.Unlock()
-
+func (c *CaptchaManager) Serve() (err error) {
 	// Cleanup procedure
-	ticker := time.NewTicker(time.Second)
+	cleanuupTicker := time.NewTicker(time.Second)
+	loadTicker := time.NewTicker(time.Minute)
 	for {
 		select {
-		case <-ctx.Done():
-			ticker.Stop()
+		case <-c.ctx.Done():
+			cleanuupTicker.Stop()
 			return
-		case <-ticker.C:
+		case <-cleanuupTicker.C:
 			c.cleanup()
+		case <-loadTicker.C:
+			c.load()
 		}
 	}
 }
 
-func (c *ClassificationCaptchaManager) GetTask() *CaptchaTask {
-	taskTemplateIndex := rand.Intn(len(c.temlates))
-	template := c.temlates[taskTemplateIndex]
+func (c *CaptchaManager) Task() *Challenge {
+	challengeIdx := rand.Intn(len(c.templates))
+	challengeTemplate := c.templates[challengeIdx]
 
-	task := CaptchaTask{
-		Description: template.Description,
-		Images:      make([]string, c.complexity),
-		Complexity:  c.complexity,
-		ts:          time.Now(),
+	challenge := Challenge{
+		ChallengeTemplate: ChallengeTemplate{
+			Description: challengeTemplate.Description,
+			Images:      make([]string, c.complexity),
+		},
+		Complexity: c.complexity,
+		ts:         time.Now(),
 	}
-
 	shuffledIndex := rand.Perm(c.complexity)
-	task.Solution = slices.Clone(shuffledIndex[:c.complexity/2])
-	slices.Sort(task.Solution)
+	challenge.Solution = slices.Clone(shuffledIndex[:c.complexity/2])
+	slices.Sort(challenge.Solution)
 	for i := 0; i < c.complexity; i++ {
 		index := shuffledIndex[i]
 		if i < c.complexity/2 {
-			task.Images[index] = template.Images[index]
-
+			imageKey := fmt.Sprintf("%d:%d", challengeIdx, index)
+			challenge.Images[index] = c.imageBytes[imageKey]
 		} else {
 			for {
-				templateIndex := rand.Intn(len(c.temlates))
-				if templateIndex != taskTemplateIndex {
-					task.Images[index] = c.temlates[templateIndex].Images[index]
+				templateIndex := rand.Intn(len(c.templates))
+				if templateIndex != challengeIdx {
+					imageKey := fmt.Sprintf("%d:%d", templateIndex, index)
+					challenge.Images[index] = c.imageBytes[imageKey] //c.templates[templateIndex].Images[index]
 					break
 				}
 			}
@@ -106,27 +107,16 @@ func (c *ClassificationCaptchaManager) GetTask() *CaptchaTask {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	for {
-		task.Id = rand.Uint32()
-		if _, exists := c.tasks[task.Id]; !exists {
-			c.tasks[task.Id] = &task
+		challenge.Id = rand.Uint32()
+		if _, exists := c.tasks[challenge.Id]; !exists {
+			c.tasks[challenge.Id] = &challenge
 			break
 		}
 	}
-	return &task
+	return &challenge
 }
 
-func (c *ClassificationCaptchaManager) ResolveFile(taskId uint32, imageId int) (file string, exists bool) {
-	task, exists := c.tasks[taskId]
-	if !exists {
-		return
-	}
-	if imageId < 0 || imageId >= len(task.Images) {
-		return
-	}
-	return task.Images[imageId], true
-}
-
-func (c *ClassificationCaptchaManager) Verify(solution *CaptchaSolution) bool {
+func (c *CaptchaManager) Verify(solution *Solution) bool {
 	task, exists := c.tasks[solution.Id]
 	if !exists {
 		return false
@@ -138,10 +128,55 @@ func (c *ClassificationCaptchaManager) Verify(solution *CaptchaSolution) bool {
 	return isValid
 }
 
-func NewClassificationCaptchaManager(complexity int) *ClassificationCaptchaManager {
-	manager := ClassificationCaptchaManager{
-		complexity: complexity,
-		tasks:      make(map[uint32]*CaptchaTask),
+func (c *CaptchaManager) load() (err error) {
+	// Load configuration from file
+	content, err := os.ReadFile("/etc/aegis/captcha.json")
+	if err != nil {
+		return
 	}
+	var configuration Configuration
+	err = json.Unmarshal(content, &configuration)
+	if err != nil {
+		return
+	}
+	// Load images
+	buf := bytes.Buffer{}
+	images := map[string]string{}
+	for templateId, template := range configuration.Templates {
+		for imageId, imageFile := range template.Images {
+			var f *os.File
+			f, err = os.Open(imageFile)
+			if err != nil {
+				return
+			}
+			defer f.Close()
+
+			img, _ := jpeg.Decode(f)
+			noisedImage := addUniformNoise(img, 20)
+			buf.Reset()
+			err = jpeg.Encode(&buf, noisedImage, &jpeg.Options{Quality: 50})
+			if err != nil {
+				return
+			}
+			images[fmt.Sprintf("%d:%d", templateId, imageId)] = base64.StdEncoding.EncodeToString(buf.Bytes())
+		}
+	}
+	// Update current images
+	c.mu.Lock()
+	c.templates = configuration.Templates
+	c.imageBytes = images
+	defer c.mu.Unlock()
+	return
+}
+
+func NewClassificationCaptchaManager(ctx context.Context, complexity int) *CaptchaManager {
+	manager := CaptchaManager{
+		ctx:        ctx,
+		complexity: complexity,
+		tasks:      make(map[uint32]*Challenge),
+		imageBytes: make(map[string]string),
+	}
+	manager.load()
+	go manager.Serve()
 	return &manager
 }